Rewrite of nhflow printer like for ecb8ef19
