+++ src/nhflow_header.h
@@ -29,8 +29,7 @@ Author: Hans Bihs
 #include"nhflow_fsf_v.h"
 
 #include"nhflow_printer.h"
-#include"nhflow_vtu3D.h"
-#include"nhflow_vts3D.h"
+#include"printer_nhflow.h"
 #include"nhflow_timestep.h"
 #include"nhflow_momentum.h"
 #include"nhflow_turbulence.h"
