+++ src/printer_nhflow.h
@@ -20,8 +20,12 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 Author: Hans Bihs
 --------------------------------------------------------------------*/
 
-#include"nhflow_printer.h"
-#include"increment.h"
+#ifndef PRINTER_NHFLOW__H_
+#define PRINTER_NHFLOW__H_
+
+#include "nhflow_printer.h"
+#include "increment.h"
+#include "vtks.h"
 
 class fdm_nhf;
 class force_ale;
@@ -39,32 +43,22 @@ class nhflow_vel_probe;
 class nhflow_vel_probe_theory;
 class ioflow;
 
-#ifndef NHFLOW_VTS3D_H_
-#define NHFLOW_VTS3D_H_
-
 using namespace std;
 
-class nhflow_vts3D : public nhflow_printer, public increment
+class printer_nhflow : public nhflow_printer, public increment
 {
 
 public:
-	nhflow_vts3D(lexer*,fdm_nhf*,ghostcell*);
-	virtual ~nhflow_vts3D();
+	printer_nhflow(lexer*,fdm_nhf*,ghostcell*);
+	virtual ~printer_nhflow();
 	virtual void start(lexer*,fdm_nhf*,ghostcell*,ioflow*);
     virtual void print_vtk(lexer*,fdm_nhf*,ghostcell*);
     virtual void print_stop(lexer*,fdm_nhf*,ghostcell*,ioflow*);
     
 private:
-    void pvts(lexer*,ghostcell*);
-    void name_iter(lexer*,ghostcell*);
-    void name_time(lexer*,ghostcell*);
-    void piecename(lexer*,ghostcell*, int);
-    void extent(lexer*,int);
-    void fextent(lexer*);
+    void parallelData(lexer*,ghostcell*);
 
-    char name[200],pname[200],epsvar[200],pextent[20];
-    int iextent[6];
-    int *piextent;
+    char name[200],pname[200],epsvar[200];
     int n,iin,offset[200];
     float ffn;
     int jj;
@@ -90,6 +84,8 @@ private:
 	force_ale **pforce_ale;
     nhflow_vel_probe *pvel;
     nhflow_vel_probe_theory *pveltheo;
+
+    vtk3D *outputFormat;
 };
 
 #endif
